refactor: protocol -> dto/response 패키지 변경
# Conflicts:
#	src/main/java/kr/flab/ottsharing/service/MoneyService.java
