+++ src/main/java/kr/flab/ottsharing/dto/response/MyInfo.java
@@ -1,4 +1,4 @@
-package kr.flab.ottsharing.protocol;
+package kr.flab.ottsharing.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
