+++ src/main/java/kr/flab/ottsharing/dto/response/MyPageUpdateResult.java
@@ -1,4 +1,4 @@
-package kr.flab.ottsharing.protocol;
+package kr.flab.ottsharing.dto.response;
 
 public enum MyPageUpdateResult {
     CHANGE_COMPLETE, DUPLICATED_EMAIL, NOTHING_CHANGED
