+++ src/main/java/kr/flab/ottsharing/dto/response/MyParty.java
@@ -1,4 +1,4 @@
-package kr.flab.ottsharing.protocol;
+package kr.flab.ottsharing.dto.response;
 
 import java.util.List;
