+++ src/main/java/kr/flab/ottsharing/dto/response/PartyCreateResult.java
@@ -1,4 +1,4 @@
-package kr.flab.ottsharing.protocol;
+package kr.flab.ottsharing.dto.response;
 
 public enum PartyCreateResult {
     NOT_ENOUGH_MONEY, SUCCESS
