+++ src/main/java/kr/flab/ottsharing/dto/response/PartyJoinResult.java
@@ -1,4 +1,4 @@
-package kr.flab.ottsharing.protocol;
+package kr.flab.ottsharing.dto.response;
 
 public enum PartyJoinResult {
     SUCCESS, NOT_ENOUGH_MONEY, ON_WAITING
