+++ src/main/java/kr/flab/ottsharing/dto/response/PayResult.java
@@ -1,4 +1,4 @@
-package kr.flab.ottsharing.protocol;
+package kr.flab.ottsharing.dto.response;
 
 public enum PayResult {
     SUCCESS, NOT_ENOUGH_MONEY
