+++ src/main/java/kr/flab/ottsharing/dto/response/RegisterResult.java
@@ -1,4 +1,4 @@
-package kr.flab.ottsharing.protocol;
+package kr.flab.ottsharing.dto.response;
 
 public enum RegisterResult {
     DUPLICATE_USER_ID, DUPLICATE_EMAIL, INVALID_EMAIL, WEAK_PASSWORD, SUCCESS
