+++ src/main/java/kr/flab/ottsharing/dto/response/UserDeleteResult.java
@@ -1,4 +1,4 @@
-package kr.flab.ottsharing.protocol;
+package kr.flab.ottsharing.dto.response;
 
 public enum UserDeleteResult {
     HAS_MONEY, HAS_PARTY, SUCCESS
