+++ src/main/java/kr/flab/ottsharing/service/PartyService.java
@@ -9,17 +9,17 @@ import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
+import kr.flab.ottsharing.dto.response.MyParty;
+import kr.flab.ottsharing.dto.response.PartyCreateResult;
+import kr.flab.ottsharing.dto.response.PartyDeleteResult;
+import kr.flab.ottsharing.dto.response.PartyJoinResult;
+import kr.flab.ottsharing.dto.response.PartyMemberInfo;
+import kr.flab.ottsharing.dto.response.PayResult;
+import kr.flab.ottsharing.dto.response.UpdatePartyInfo;
 import kr.flab.ottsharing.entity.Party;
 import kr.flab.ottsharing.entity.PartyMember;
 import kr.flab.ottsharing.entity.User;
 import kr.flab.ottsharing.exception.WrongInfoException;
-import kr.flab.ottsharing.protocol.MyParty;
-import kr.flab.ottsharing.protocol.PartyCreateResult;
-import kr.flab.ottsharing.protocol.PartyDeleteResult;
-import kr.flab.ottsharing.protocol.PartyMemberInfo;
-import kr.flab.ottsharing.protocol.PartyJoinResult;
-import kr.flab.ottsharing.protocol.PayResult;
-import kr.flab.ottsharing.protocol.UpdatePartyInfo;
 import kr.flab.ottsharing.repository.PartyMemberRepository;
 import kr.flab.ottsharing.repository.PartyRepository;
 import kr.flab.ottsharing.repository.PartyWaitingRepository;
